requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.6
django-import-export==4.3.7
django-multiselectfield==0.1.13
django-widget-tweaks==1.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.1.0
psycopg2==2.9.10
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
whitenoise==6.9.0
